[Refactor] Factor out TLS certificate fetching code
