+++ components/core/Cargo.toml
@@ -22,8 +22,10 @@ hex = "*"
 lazy_static = "*"
 libc = "0.2.76"
 log = "*"
+native-tls = { version = "*", features = ["vendored"] }
 os_info = "*"
 paste = "1.0"
+pem = "*"
 rand = "*"
 regex = "*"
 rust-crypto = "*"
