+++ components/core/build.rs
@@ -1,9 +1,11 @@
+use std::{env,
+          fs,
+          path::Path};
+
 #[cfg(windows)]
 fn main() {
-    use std::{env,
-              fs::File,
-              io::prelude::*,
-              path::Path};
+    use std::{fs::File,
+              io::prelude::*};
 
     cc::Build::new().file("./src/os/users/admincheck.c")
                     .compile("libadmincheck.a");
@@ -12,7 +14,17 @@ fn main() {
     if let Ok(key) = env::var("HAB_CRYPTO_KEY") {
         file.write_all(&base64::decode(&key).unwrap()).unwrap();
     }
+
+    populate_cacert();
 }
 
 #[cfg(not(windows))]
-fn main() {}
+fn main() { populate_cacert(); }
+
+pub fn populate_cacert() {
+    let src = env::var("SSL_CERT_FILE").unwrap();
+    let dst = Path::new(&env::var("OUT_DIR").unwrap()).join("cacert.pem");
+    if !dst.exists() {
+        fs::copy(src, dst).unwrap();
+    }
+}
