+++ components/core/src/lib.rs
@@ -17,6 +17,7 @@ pub mod origin;
 pub mod os;
 pub mod package;
 pub mod service;
+pub mod tls;
 pub mod url;
 pub mod util;
