+++ components/http-client/src/lib.rs
@@ -4,7 +4,6 @@ extern crate log;
 mod api_client;
 mod error;
 
-pub use crate::{api_client::{certificates,
-                             ApiClient},
+pub use crate::{api_client::ApiClient,
                 error::{Error,
                         Result}};
