+++ components/sup/src/event/error.rs
@@ -11,7 +11,7 @@ pub type Result<T> = result::Result<T, Error>;
 #[derive(Debug)]
 pub enum Error {
     ConnectNatsServer,
-    HabitatHttpClient(habitat_http_client::Error),
+    HabitatCore(habitat_core::Error),
     NativeTls(native_tls::Error),
     Rants(RantsError),
 }
@@ -28,7 +28,7 @@ impl fmt::Display for Error {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         match self {
             Error::ConnectNatsServer => "Could not establish connection to NATS server".fmt(f),
-            Error::HabitatHttpClient(_) => "{}".fmt(f),
+            Error::HabitatCore(_) => "{}".fmt(f),
             Error::NativeTls(e) => format!("{}", e).fmt(f),
             Error::Rants(e) => format!("{}", e).fmt(f),
         }
@@ -39,15 +39,15 @@ impl error::Error for Error {
     fn source(&self) -> Option<&(dyn error::Error + 'static)> {
         match self {
             Error::ConnectNatsServer => None,
-            Error::HabitatHttpClient(ref e) => Some(e),
+            Error::HabitatCore(ref e) => Some(e),
             Error::Rants(ref e) => Some(e),
             Error::NativeTls(ref e) => Some(e),
         }
     }
 }
 
-impl From<habitat_http_client::Error> for Error {
-    fn from(error: habitat_http_client::Error) -> Self { Error::HabitatHttpClient(error) }
+impl From<habitat_core::Error> for Error {
+    fn from(error: habitat_core::Error) -> Self { Error::HabitatCore(error) }
 }
 
 impl From<RantsError> for Error {
